+++ examples/SX126x_api_tx/SX126x_api_tx.ino
@@ -111,7 +111,7 @@ void settingFunction() {
   Serial.print("Set frequency to ");
   Serial.print(rfFrequency / 1000000);
   Serial.println(" Mhz");
-  uint32_t rfFreq = rfFrequency * 33554432UL / 32000000UL;
+  uint32_t rfFreq = ((uint64_t) rfFrequency * 33554432UL) / 32000000UL;
   Api.setRfFrequency(rfFreq);
 
   // Set tx power to selected TX power
